eCAL Core: Create a new eCAL::types target that contains eCAL type definitions, but doesn't require linking to eCAL::core or other libraries (e.g. like protobuf)
