+++ testing/ecal/types_test/CMakeLists.txt
@@ -1,0 +1,34 @@
+# ========================= eCAL LICENSE =================================
+#
+# Copyright (C) 2016 - 2019 Continental Corporation
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+# 
+#      http://www.apache.org/licenses/LICENSE-2.0
+# 
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# ========================= eCAL LICENSE =================================
+
+project(test_types)
+
+set(types_test_src
+  src/types_test.cpp
+)
+
+add_executable(${PROJECT_NAME} ${types_test_src})
+
+target_link_libraries(${PROJECT_NAME}
+  PRIVATE
+    eCAL::types
+)
+
+target_compile_features(${PROJECT_NAME} PRIVATE cxx_std_14)
+
+set_property(TARGET ${PROJECT_NAME} PROPERTY FOLDER testing/ecal/core)
